IGNITE-17167 ConfigurationAsmGenerator decomposition. Part 2, DynamicConfiguration-related classes.
