+++ modules/configuration/src/main/java/org/apache/ignite/internal/configuration/asm/ConfigurationAsmGenerator.java
@@ -17,47 +17,27 @@
 
 package org.apache.ignite.internal.configuration.asm;
 
-import static com.facebook.presto.bytecode.Access.FINAL;
-import static com.facebook.presto.bytecode.Access.PRIVATE;
-import static com.facebook.presto.bytecode.Access.PUBLIC;
-import static com.facebook.presto.bytecode.Access.STATIC;
-import static com.facebook.presto.bytecode.Access.SYNTHETIC;
-import static com.facebook.presto.bytecode.Parameter.arg;
 import static com.facebook.presto.bytecode.ParameterizedType.type;
 import static com.facebook.presto.bytecode.ParameterizedType.typeFromJavaClassName;
-import static com.facebook.presto.bytecode.expression.BytecodeExpressions.constantBoolean;
-import static com.facebook.presto.bytecode.expression.BytecodeExpressions.constantClass;
-import static com.facebook.presto.bytecode.expression.BytecodeExpressions.constantInt;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.constantNull;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.constantString;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.inlineIf;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.invokeDynamic;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.isNull;
-import static com.facebook.presto.bytecode.expression.BytecodeExpressions.newArray;
 import static com.facebook.presto.bytecode.expression.BytecodeExpressions.newInstance;
-import static com.facebook.presto.bytecode.expression.BytecodeExpressions.set;
 import static java.lang.invoke.MethodType.methodType;
 import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
-import static java.util.EnumSet.of;
-import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toCollection;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
-import static org.apache.ignite.internal.configuration.asm.DirectProxyAsmGenerator.newDirectProxyLambda;
 import static org.apache.ignite.internal.configuration.asm.SchemaClassesInfo.changeClassName;
 import static org.apache.ignite.internal.configuration.asm.SchemaClassesInfo.configurationClassName;
-import static org.apache.ignite.internal.configuration.asm.SchemaClassesInfo.nodeClassName;
 import static org.apache.ignite.internal.configuration.asm.SchemaClassesInfo.viewClassName;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.extensionsFields;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isConfigValue;
-import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isInjectedName;
-import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isInternalId;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isNamedConfigValue;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isPolymorphicConfig;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isPolymorphicConfigInstance;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isPolymorphicId;
-import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.isValue;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.polymorphicInstanceId;
 import static org.apache.ignite.internal.configuration.util.ConfigurationUtil.schemaFields;
 import static org.apache.ignite.internal.util.ArrayUtils.nullOrEmpty;
@@ -73,7 +53,6 @@ import com.facebook.presto.bytecode.ClassGenerator;
 import com.facebook.presto.bytecode.FieldDefinition;
 import com.facebook.presto.bytecode.MethodDefinition;
 import com.facebook.presto.bytecode.ParameterizedType;
-import com.facebook.presto.bytecode.Variable;
 import com.facebook.presto.bytecode.expression.BytecodeExpression;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
@@ -88,14 +67,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Queue;
 import java.util.Set;
-import java.util.function.BiFunction;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
-import org.apache.ignite.configuration.ConfigurationProperty;
-import org.apache.ignite.configuration.ConfigurationTree;
-import org.apache.ignite.configuration.ConfigurationValue;
 import org.apache.ignite.configuration.ConfigurationWrongPolymorphicTypeIdException;
-import org.apache.ignite.configuration.NamedConfigurationTree;
 import org.apache.ignite.configuration.RootKey;
 import org.apache.ignite.configuration.annotation.AbstractConfiguration;
 import org.apache.ignite.configuration.annotation.Config;
@@ -105,21 +79,15 @@ import org.apache.ignite.configuration.annotation.NamedConfigValue;
 import org.apache.ignite.configuration.annotation.PolymorphicConfig;
 import org.apache.ignite.configuration.annotation.PolymorphicConfigInstance;
 import org.apache.ignite.configuration.annotation.PolymorphicId;
-import org.apache.ignite.internal.configuration.ConfigurationNode;
-import org.apache.ignite.internal.configuration.ConfigurationTreeWrapper;
 import org.apache.ignite.internal.configuration.DynamicConfiguration;
 import org.apache.ignite.internal.configuration.DynamicConfigurationChanger;
-import org.apache.ignite.internal.configuration.DynamicProperty;
-import org.apache.ignite.internal.configuration.NamedListConfiguration;
 import org.apache.ignite.internal.configuration.TypeUtils;
-import org.apache.ignite.internal.configuration.direct.DirectPropertyProxy;
 import org.apache.ignite.internal.configuration.tree.InnerNode;
 import org.apache.ignite.internal.configuration.tree.NamedListNode;
 import org.apache.ignite.internal.configuration.util.ConfigurationUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.objectweb.asm.Handle;
-import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
 
 /**
@@ -275,7 +243,8 @@ public class ConfigurationAsmGenerator extends Methods {
                     internalIdField
             ).generate());
 
-            ClassDefinition cfgImplClassDef = createCfgImplClass(
+            classDefs.addAll(new CfgImplAsmGenerator(
+                    this,
                     schemaClass,
                     internalExtensions,
                     polymorphicExtensions,
@@ -283,25 +252,7 @@ public class ConfigurationAsmGenerator extends Methods {
                     internalExtensionsFields,
                     polymorphicExtensionsFields,
                     internalIdField
-            );
-
-            classDefs.add(cfgImplClassDef);
-
-            for (Class<?> polymorphicExtension : polymorphicExtensions) {
-                // Only the fields of a specific instance of a polymorphic configuration.
-                Collection<Field> polymorphicFields = polymorphicExtensionsFields.stream()
-                        .filter(f -> f.getDeclaringClass() == polymorphicExtension)
-                        .collect(toList());
-
-                classDefs.add(createPolymorphicExtensionCfgImplClass(
-                        schemaClass,
-                        polymorphicExtension,
-                        cfgImplClassDef,
-                        schemaFields,
-                        polymorphicFields,
-                        internalIdField
-                ));
-            }
+            ).generate());
 
             ClassDefinition directProxyClassDef = new DirectProxyAsmGenerator(
                     this,
@@ -422,464 +373,6 @@ public class ConfigurationAsmGenerator extends Methods {
     }
 
     /**
-     * Construct a {@link DynamicConfiguration} definition for a configuration schema.
-     *
-     * @param schemaClass        Configuration schema class.
-     * @param internalExtensions Internal extensions of the configuration schema.
-     * @param schemaFields       Fields of the schema class.
-     * @param internalFields     Fields of internal extensions of the configuration schema.
-     * @param polymorphicFields  Fields of polymorphic extensions of the configuration schema.
-     * @param internalIdField    Internal id field or {@code null} if it's not present.
-     * @return Constructed {@link DynamicConfiguration} definition for the configuration schema.
-     */
-    private ClassDefinition createCfgImplClass(
-            Class<?> schemaClass,
-            Set<Class<?>> internalExtensions,
-            Set<Class<?>> polymorphicExtensions,
-            Collection<Field> schemaFields,
-            Collection<Field> internalFields,
-            Collection<Field> polymorphicFields,
-            @Nullable Field internalIdField
-    ) {
-        SchemaClassesInfo schemaClassInfo = schemasInfo.get(schemaClass);
-
-        // Configuration impl class definition.
-        ClassDefinition classDef = new ClassDefinition(
-                of(PUBLIC, FINAL),
-                internalName(schemaClassInfo.cfgImplClassName),
-                type(DynamicConfiguration.class),
-                configClassInterfaces(schemaClass, internalExtensions)
-        );
-
-        // Fields.
-        Map<String, FieldDefinition> fieldDefs = new HashMap<>();
-
-        // To store the id of the polymorphic configuration instance.
-        FieldDefinition polymorphicTypeIdFieldDef = null;
-
-        for (Field schemaField : concat(schemaFields, internalFields, polymorphicFields)) {
-            String fieldName = fieldName(schemaField);
-
-            FieldDefinition fieldDef = addConfigurationImplField(classDef, schemaField, fieldName);
-
-            fieldDefs.put(fieldName, fieldDef);
-
-            if (isPolymorphicId(schemaField)) {
-                polymorphicTypeIdFieldDef = fieldDef;
-            }
-        }
-
-        if (internalIdField != null) {
-            // Internal id dynamic property is stored as a regular field.
-            String fieldName = internalIdField.getName();
-
-            FieldDefinition fieldDef = addConfigurationImplField(classDef, internalIdField, fieldName);
-
-            fieldDefs.put(fieldName, fieldDef);
-        }
-
-        FieldDefinition internalConfigTypesFieldDef = null;
-
-        if (!internalExtensions.isEmpty()) {
-            internalConfigTypesFieldDef = classDef.declareField(
-                    of(PRIVATE, FINAL),
-                    INTERNAL_CONFIG_TYPES_FIELD_NAME,
-                    Class[].class
-            );
-        }
-
-        // Constructor
-        addConfigurationImplConstructor(
-                classDef,
-                schemaClass,
-                internalExtensions,
-                fieldDefs,
-                schemaFields,
-                internalFields,
-                polymorphicFields,
-                internalIdField,
-                internalConfigTypesFieldDef
-        );
-
-        // org.apache.ignite.internal.configuration.DynamicProperty#directProxy
-        addDirectProxyMethod(schemaClassInfo, classDef);
-
-        // Getter for the internal id.
-        if (internalIdField != null) {
-            addConfigurationImplGetMethod(classDef, internalIdField, fieldDefs.get(internalIdField.getName()));
-        }
-
-        for (Field schemaField : concat(schemaFields, internalFields)) {
-            addConfigurationImplGetMethod(classDef, schemaField, fieldDefs.get(fieldName(schemaField)));
-        }
-
-        // org.apache.ignite.internal.configuration.DynamicConfiguration#configType
-        addCfgImplConfigTypeMethod(classDef, typeFromJavaClassName(schemaClassInfo.cfgClassName));
-
-        if (internalConfigTypesFieldDef != null) {
-            addCfgImplInternalConfigTypesMethod(classDef, internalConfigTypesFieldDef);
-        }
-
-        if (!polymorphicExtensions.isEmpty()) {
-            addCfgSpecificConfigTreeMethod(classDef, schemaClass, polymorphicExtensions, polymorphicTypeIdFieldDef);
-
-            addCfgRemoveMembersMethod(
-                    classDef,
-                    schemaClass,
-                    polymorphicExtensions,
-                    fieldDefs,
-                    polymorphicFields,
-                    polymorphicTypeIdFieldDef
-            );
-
-            addCfgAddMembersMethod(
-                    classDef,
-                    schemaClass,
-                    polymorphicExtensions,
-                    fieldDefs,
-                    polymorphicFields,
-                    polymorphicTypeIdFieldDef
-            );
-
-            addCfgImplPolymorphicInstanceConfigTypeMethod(
-                    classDef,
-                    schemaClass,
-                    polymorphicExtensions,
-                    polymorphicTypeIdFieldDef
-            );
-        }
-
-        return classDef;
-    }
-
-    /**
-     * Declares field that corresponds to configuration value. Depending on the schema, 3 options possible:
-     * <ul>
-     *     <li>
-     *         {@code @Value public type fieldName}<br/>becomes<br/>
-     *         {@code public DynamicProperty fieldName}
-     *     </li>
-     *     <li>
-     *         {@code @ConfigValue public MyConfigurationSchema fieldName}<br/>becomes<br/>
-     *         {@code public MyConfiguration fieldName}
-     *     </li>
-     *     <li>
-     *         {@code @NamedConfigValue public type fieldName}<br/>becomes<br/>
-     *         {@code public NamedListConfiguration fieldName}
-     *     </li>
-     *     <li>
-     *         {@code @PolymorphicId public String fieldName}<br/>becomes<br/>
-     *         {@code public String fieldName}
-     *     </li>
-     * </ul>
-     *
-     * @param classDef    Configuration impl class definition.
-     * @param schemaField Configuration Schema class field.
-     * @param fieldName   Field name, if {@code null} will be used {@link Field#getName}.
-     * @return Declared field definition.
-     */
-    private FieldDefinition addConfigurationImplField(
-            ClassDefinition classDef,
-            Field schemaField,
-            String fieldName
-    ) {
-        ParameterizedType fieldType;
-
-        if (isConfigValue(schemaField)) {
-            fieldType = typeFromJavaClassName(schemasInfo.get(schemaField.getType()).cfgImplClassName);
-        } else if (isNamedConfigValue(schemaField)) {
-            fieldType = type(NamedListConfiguration.class);
-        } else {
-            fieldType = type(DynamicProperty.class);
-        }
-
-        return classDef.declareField(of(PUBLIC), fieldName, fieldType);
-    }
-
-    /**
-     * Implements default constructor for the configuration class. It initializes all fields and adds them to members collection.
-     *
-     * @param classDef Configuration impl class definition.
-     * @param schemaClass Configuration schema class.
-     * @param internalExtensions Internal extensions of the configuration schema.
-     * @param fieldDefs Field definitions for all fields of configuration impl class.
-     * @param schemaFields Fields of the schema class.
-     * @param internalFields Fields of internal extensions of the configuration schema.
-     * @param polymorphicFields Fields of polymorphic extensions of the configuration schema.
-     * @param internalIdField Internal id field or {@code null} if it's not present.
-     * @param internalConfigTypesFieldDef Field definition for {@link DynamicConfiguration#internalConfigTypes},
-     *      {@code null} if there are no internal extensions.
-     */
-    private void addConfigurationImplConstructor(
-            ClassDefinition classDef,
-            Class<?> schemaClass,
-            Set<Class<?>> internalExtensions,
-            Map<String, FieldDefinition> fieldDefs,
-            Collection<Field> schemaFields,
-            Collection<Field> internalFields,
-            Collection<Field> polymorphicFields,
-            @Nullable Field internalIdField,
-            @Nullable FieldDefinition internalConfigTypesFieldDef
-    ) {
-        MethodDefinition ctor = classDef.declareConstructor(
-                of(PUBLIC),
-                arg("prefix", List.class),
-                arg("key", String.class),
-                arg("rootKey", RootKey.class),
-                arg("changer", DynamicConfigurationChanger.class),
-                arg("listenOnly", boolean.class)
-        );
-
-        Variable rootKeyVar = ctor.getScope().getVariable("rootKey");
-        Variable changerVar = ctor.getScope().getVariable("changer");
-        Variable listenOnlyVar = ctor.getScope().getVariable("listenOnly");
-
-        SchemaClassesInfo schemaClassInfo = schemasInfo.get(schemaClass);
-
-        Variable thisVar = ctor.getThis();
-
-        BytecodeBlock ctorBody = ctor.getBody()
-                .append(thisVar)
-                .append(ctor.getScope().getVariable("prefix"))
-                .append(ctor.getScope().getVariable("key"))
-                .append(rootKeyVar)
-                .append(changerVar)
-                .append(listenOnlyVar)
-                .invokeConstructor(DYNAMIC_CONFIGURATION_CTOR);
-
-        BytecodeExpression thisKeysVar = thisVar.getField("keys", List.class);
-
-        // Wrap object into list to reuse the loop below.
-        List<Field> internalIdFieldAsList = internalIdField == null ? emptyList() : List.of(internalIdField);
-
-        int newIdx = 0;
-        for (Field schemaField : concat(schemaFields, internalFields, polymorphicFields, internalIdFieldAsList)) {
-            String fieldName = schemaField.getName();
-
-            BytecodeExpression newValue;
-
-            if (isValue(schemaField) || isPolymorphicId(schemaField) || isInjectedName(schemaField) || isInternalId(schemaField)) {
-                // A field with @InjectedName is special (auxiliary), it is not stored in storages as a regular field, and therefore there
-                // is no direct access to it. It is stored in the InnerNode and does not participate in its traversal, so in order to get
-                // it we need to get the InnerNode, and only then the value of this field.
-
-                // newValue = new DynamicProperty(this.keys, fieldName, rootKey, changer, listenOnly, readOnly);
-                newValue = newInstance(
-                        DynamicProperty.class,
-                        thisKeysVar,
-                        constantString(isInjectedName(schemaField) ? InnerNode.INJECTED_NAME
-                                : isInternalId(schemaField) ? InnerNode.INTERNAL_ID : schemaField.getName()),
-                        rootKeyVar,
-                        changerVar,
-                        listenOnlyVar,
-                        constantBoolean(isPolymorphicId(schemaField) || isInjectedName(schemaField) || isInternalId(schemaField))
-                );
-            } else {
-                SchemaClassesInfo fieldInfo = schemasInfo.get(schemaField.getType());
-
-                ParameterizedType cfgImplParameterizedType = typeFromJavaClassName(fieldInfo.cfgImplClassName);
-
-                if (isConfigValue(schemaField)) {
-                    // newValue = new MyConfigurationImpl(super.keys, fieldName, rootKey, changer, listenOnly);
-                    newValue = newInstance(
-                            cfgImplParameterizedType,
-                            thisKeysVar,
-                            constantString(fieldName),
-                            rootKeyVar,
-                            changerVar,
-                            listenOnlyVar
-                    );
-                } else {
-                    // We have to create method "$new$<idx>" to reference it in lambda expression. That's the way it
-                    // works, it'll invoke constructor with all 5 arguments, not just 2 as in BiFunction.
-                    MethodDefinition newMtd = classDef.declareMethod(
-                            of(PRIVATE, STATIC, SYNTHETIC),
-                            "$new$" + newIdx++,
-                            typeFromJavaClassName(fieldInfo.cfgClassName),
-                            arg("rootKey", RootKey.class),
-                            arg("changer", DynamicConfigurationChanger.class),
-                            arg("listenOnly", boolean.class),
-                            arg("prefix", List.class),
-                            arg("key", String.class)
-                    );
-
-                    // newValue = new NamedListConfiguration(this.keys, fieldName, rootKey, changer, listenOnly,
-                    //      (p, k) -> new ValueConfigurationImpl(p, k, rootKey, changer, listenOnly),
-                    //      (p, c) -> new ValueDirectProxy(p, c),
-                    //      new ValueConfigurationImpl(this.keys, "any", rootKey, changer, true)
-                    // );
-                    newValue = newInstance(
-                            NamedListConfiguration.class,
-                            thisKeysVar,
-                            constantString(fieldName),
-                            rootKeyVar,
-                            changerVar,
-                            listenOnlyVar,
-                            invokeDynamic(
-                                    LAMBDA_METAFACTORY,
-                                    asList(
-                                            getMethodType(getType(Object.class), getType(Object.class), getType(Object.class)),
-                                            new Handle(
-                                                    Opcodes.H_INVOKESTATIC,
-                                                    internalName(schemaClassInfo.cfgImplClassName),
-                                                    newMtd.getName(),
-                                                    newMtd.getMethodDescriptor(),
-                                                    false
-                                            ),
-                                            getMethodType(
-                                                    typeFromJavaClassName(fieldInfo.cfgClassName).getAsmType(),
-                                                    getType(List.class),
-                                                    getType(String.class)
-                                            )
-                                    ),
-                                    "apply",
-                                    BiFunction.class,
-                                    rootKeyVar,
-                                    changerVar,
-                                    listenOnlyVar
-                            ),
-                            newDirectProxyLambda(fieldInfo),
-                            newInstance(
-                                    cfgImplParameterizedType,
-                                    thisKeysVar,
-                                    constantString("any"),
-                                    rootKeyVar,
-                                    changerVar,
-                                    constantBoolean(true)
-                            ).cast(ConfigurationProperty.class)
-                    );
-
-                    newMtd.getBody()
-                            .append(newInstance(
-                                    cfgImplParameterizedType,
-                                    newMtd.getScope().getVariable("prefix"),
-                                    newMtd.getScope().getVariable("key"),
-                                    newMtd.getScope().getVariable("rootKey"),
-                                    newMtd.getScope().getVariable("changer"),
-                                    newMtd.getScope().getVariable("listenOnly")
-                            ))
-                            .retObject();
-                }
-            }
-
-            FieldDefinition fieldDef = fieldDefs.get(fieldName(schemaField));
-
-            // this.field = newValue;
-            ctorBody.append(thisVar.setField(fieldDef, newValue));
-
-            if (!isPolymorphicConfigInstance(schemaField.getDeclaringClass()) && !isInternalId(schemaField)) {
-                // add(this.field);
-                ctorBody.append(thisVar.invoke(DYNAMIC_CONFIGURATION_ADD_MTD, thisVar.getField(fieldDef)));
-            }
-        }
-
-        if (internalConfigTypesFieldDef != null) {
-            assert !internalExtensions.isEmpty() : classDef;
-
-            // Class[] tmp;
-            Variable tmpVar = ctor.getScope().createTempVariable(Class[].class);
-
-            BytecodeBlock initInternalConfigTypesField = new BytecodeBlock();
-
-            // tmp = new Class[size];
-            initInternalConfigTypesField.append(tmpVar.set(newArray(type(Class[].class), internalExtensions.size())));
-
-            int i = 0;
-
-            for (Class<?> extension : internalExtensions) {
-                // tmp[i] = InternalTableConfiguration.class;
-                initInternalConfigTypesField.append(set(
-                        tmpVar,
-                        constantInt(i++),
-                        constantClass(typeFromJavaClassName(configurationClassName(extension)))
-                ));
-            }
-
-            // this._internalConfigTypes = tmp;
-            initInternalConfigTypesField.append(setThisFieldCode(ctor, tmpVar, internalConfigTypesFieldDef));
-
-            ctorBody.append(initInternalConfigTypesField);
-        }
-
-        ctorBody.ret();
-    }
-
-    /**
-     * Generates {@link ConfigurationNode#directProxy()} method that returns new instance every time.
-     *
-     * @param schemaClassInfo Schema class info.
-     * @param classDef Class definition.
-     */
-    private void addDirectProxyMethod(
-            SchemaClassesInfo schemaClassInfo,
-            ClassDefinition classDef
-    ) {
-        MethodDefinition methodDef = classDef.declareMethod(
-                of(PUBLIC), "directProxy", type(DirectPropertyProxy.class)
-        );
-
-        methodDef.getBody().append(newInstance(
-                typeFromJavaClassName(schemaClassInfo.directProxyClassName),
-                methodDef.getThis().invoke("keyPath", List.class),
-                methodDef.getThis().getField("changer", DynamicConfigurationChanger.class)
-        ));
-
-        methodDef.getBody().retObject();
-    }
-
-    /**
-     * Implements accessor method in configuration impl class.
-     *
-     * @param classDef    Configuration impl class definition.
-     * @param schemaField Configuration Schema class field.
-     * @param fieldDefs   Field definitions.
-     */
-    private void addConfigurationImplGetMethod(
-            ClassDefinition classDef,
-            Field schemaField,
-            FieldDefinition... fieldDefs
-    ) {
-        assert !nullOrEmpty(fieldDefs);
-
-        Class<?> schemaFieldType = schemaField.getType();
-
-        String fieldName = schemaField.getName();
-
-        ParameterizedType returnType;
-
-        SchemaClassesInfo schemaClassInfo = schemasInfo.get(schemaFieldType);
-
-        if (isConfigValue(schemaField)) {
-            returnType = typeFromJavaClassName(schemaClassInfo.cfgClassName);
-        } else if (isNamedConfigValue(schemaField)) {
-            returnType = type(NamedConfigurationTree.class);
-        } else {
-            assert isValue(schemaField) || isPolymorphicId(schemaField) || isInjectedName(schemaField)
-                    || isInternalId(schemaField) : schemaField;
-
-            returnType = type(ConfigurationValue.class);
-        }
-
-        // public ConfigurationProperty fieldName()
-        MethodDefinition viewMtd = classDef.declareMethod(
-                of(PUBLIC),
-                fieldName,
-                returnType
-        );
-
-        // result = this.field;
-        BytecodeBlock body = viewMtd.getBody().append(getThisFieldCode(viewMtd, fieldDefs));
-
-        if (isPolymorphicConfig(schemaFieldType) && isConfigValue(schemaField)) {
-            // result = this.field.specificConfigTree();
-            body.invokeVirtual(SPECIFIC_CONFIG_TREE_MTD);
-        }
-
-        // return result;
-        body.retObject();
-    }
-
-    /**
      * Replaces first letter in string with its upper-cased variant.
      *
      * @param name Some string.
@@ -945,385 +438,6 @@ public class ConfigurationAsmGenerator extends Methods {
 
         return result.toArray(new ParameterizedType[0]);
     }
-
-    /**
-     * Add {@link DynamicConfiguration#configType} method implementation to the class.
-     *
-     * <p>It looks like the following code:
-     * <pre><code>
-     * public Class configType() {
-     *     return RootConfiguration.class;
-     * }
-     * </code></pre>
-     *
-     * @param classDef Class definition.
-     * @param clazz Definition of the configuration interface, for example {@code RootConfiguration}.
-     */
-    private static void addCfgImplConfigTypeMethod(ClassDefinition classDef, ParameterizedType clazz) {
-        classDef.declareMethod(of(PUBLIC), "configType", type(Class.class))
-                .getBody()
-                .append(constantClass(clazz))
-                .retObject();
-    }
-
-    /**
-     * Add {@link DynamicConfiguration#internalConfigTypes} method implementation to the class.
-     *
-     * <p>It looks like the following code:
-     * <pre><code>
-     * public Class<?>[] internalConfigTypes() {
-     *     return new Class<?>[]{FirstInternalTableConfiguration.class, SecondInternalTableConfiguration.class};
-     * }
-     * </code></pre>
-     *
-     * @param classDef Class definition.
-     * @param internalConfigTypesDef Definition of the field in which the interfaces of the internal configuration extensions are stored.
-     */
-    private static void addCfgImplInternalConfigTypesMethod(ClassDefinition classDef, FieldDefinition internalConfigTypesDef) {
-        MethodDefinition internalConfigTypesMtd = classDef.declareMethod(of(PUBLIC), "internalConfigTypes", type(Class[].class));
-
-        internalConfigTypesMtd
-                .getBody()
-                .append(getThisFieldCode(internalConfigTypesMtd, internalConfigTypesDef))
-                .retObject();
-    }
-
-    /**
-     * Add {@link DynamicConfiguration#polymorphicInstanceConfigType} method implementation to the class.
-     *
-     * <p>It looks like the following code:
-     * <pre><code>
-     * public Class polymorphicInstanceConfigType() {
-     *      InnerNode val = this.isRemovedFromNamedList() ? this.currentValue() : this.refreshValue();
-     *      String typeId = val.polymorphicTypeId;
-     *      switch(typeId) {
-     *          case "hash":
-     *              return HashIndexConfiguration.class;
-     *          case "sorted"
-     *              return SortedIndexConfiguration.class;
-     *          default:
-     *              throw new ConfigurationWrongPolymorphicTypeIdException(typeId);
-     *      }
-     * }
-     * </code></pre>
-     *
-     * @param classDef Class definition.
-     * @param schemaClass Polymorphic configuration schema (parent).
-     * @param polymorphicExtensions Polymorphic configuration instance schemas (children).
-     * @param polymorphicTypeIdFieldDef Identification field for the polymorphic configuration instance.
-     */
-    private static void addCfgImplPolymorphicInstanceConfigTypeMethod(
-            ClassDefinition classDef,
-            Class<?> schemaClass,
-            Set<Class<?>> polymorphicExtensions,
-            FieldDefinition polymorphicTypeIdFieldDef
-    ) {
-        MethodDefinition polymorphicInstanceConfigTypeMtd = classDef.declareMethod(
-                of(PUBLIC),
-                "polymorphicInstanceConfigType",
-                type(Class.class)
-        );
-
-        // String tmpStr;
-        Variable tmpStrVar = polymorphicInstanceConfigTypeMtd.getScope().createTempVariable(String.class);
-
-        StringSwitchBuilder switchBuilder = new StringSwitchBuilder(polymorphicInstanceConfigTypeMtd.getScope())
-                .expression(tmpStrVar)
-                .defaultCase(throwException(ConfigurationWrongPolymorphicTypeIdException.class, tmpStrVar));
-
-        for (Class<?> polymorphicExtension : polymorphicExtensions) {
-            switchBuilder.addCase(
-                    polymorphicInstanceId(polymorphicExtension),
-                    constantClass(typeFromJavaClassName(configurationClassName(polymorphicExtension))).ret()
-            );
-        }
-
-        // ConfigNode
-        ParameterizedType nodeType = typeFromJavaClassName(nodeClassName(schemaClass));
-
-        // Object tmpObj;
-        Variable tmpObjVar = polymorphicInstanceConfigTypeMtd.getScope().createTempVariable(Object.class);
-
-        // this;
-        Variable thisVar = polymorphicInstanceConfigTypeMtd.getThis();
-
-        // tmpObj = this.isRemovedFromNamedList() ? this.currentValue() : this.refreshValue();
-        // tmpStr = ((ConfigNode) tmpObj).typeId;
-        // switch(tmpStr) ...
-        polymorphicInstanceConfigTypeMtd.getBody()
-                .append(tmpObjVar.set(inlineIf(
-                        thisVar.invoke(IS_REMOVED_FROM_NAMED_LIST_MTD),
-                        thisVar.invoke(CURRENT_VALUE_MTD),
-                        thisVar.invoke(REFRESH_VALUE_MTD))
-                ))
-                .append(tmpStrVar.set(tmpObjVar.cast(nodeType).getField(polymorphicTypeIdFieldDef.getName(), String.class)))
-                .append(switchBuilder.build())
-                .ret();
-    }
-
-    /**
-     * Create a {@code *CfgImpl} for the polymorphic configuration instance schema.
-     *
-     * @param schemaClass           Polymorphic configuration schema (parent).
-     * @param polymorphicExtension  Polymorphic configuration instance schema (child).
-     * @param schemaCfgImplClassDef {@link DynamicConfiguration} definition for the polymorphic configuration schema {@code schemaClass}.
-     * @param schemaFields          Schema fields of polymorphic configuration {@code schemaClass}.
-     * @param polymorphicFields     Schema fields of a polymorphic configuration instance {@code polymorphicExtension}.
-     * @param internalIdField       Internal id field or {@code null} if it's not present.
-     */
-    private ClassDefinition createPolymorphicExtensionCfgImplClass(
-            Class<?> schemaClass,
-            Class<?> polymorphicExtension,
-            ClassDefinition schemaCfgImplClassDef,
-            Collection<Field> schemaFields,
-            Collection<Field> polymorphicFields,
-            @Nullable Field internalIdField
-    ) {
-        SchemaClassesInfo schemaClassInfo = schemasInfo.get(schemaClass);
-        SchemaClassesInfo polymorphicExtensionClassInfo = schemasInfo.get(polymorphicExtension);
-
-        // Configuration impl class definition.
-        ClassDefinition classDef = new ClassDefinition(
-                of(PUBLIC, FINAL),
-                internalName(polymorphicExtensionClassInfo.cfgImplClassName),
-                type(ConfigurationTreeWrapper.class),
-                configClassInterfaces(polymorphicExtension, Set.of())
-        );
-
-        // private final ParentCfgImpl this$0;
-        FieldDefinition parentCfgImplFieldDef = classDef.declareField(
-                of(PRIVATE, FINAL),
-                "this$0",
-                typeFromJavaClassName(schemaClassInfo.cfgImplClassName)
-        );
-
-        // Constructor.
-        MethodDefinition constructorMtd = classDef.declareConstructor(
-                of(PUBLIC),
-                arg("delegate", typeFromJavaClassName(schemaClassInfo.cfgImplClassName))
-        );
-
-        Variable delegateVar = constructorMtd.getScope().getVariable("delegate");
-
-        // Constructor body.
-        // super(parent);
-        // this.this$0 = parent;
-        constructorMtd.getBody()
-                .append(constructorMtd.getThis())
-                .append(delegateVar)
-                .invokeConstructor(ConfigurationTreeWrapper.class, ConfigurationTree.class)
-                .append(constructorMtd.getThis().setField(
-                        parentCfgImplFieldDef,
-                        delegateVar
-                ))
-                .ret();
-
-        Map<String, FieldDefinition> fieldDefs = schemaCfgImplClassDef.getFields().stream()
-                .collect(toMap(FieldDefinition::getName, identity()));
-
-        if (internalIdField != null) {
-            addConfigurationImplGetMethod(classDef, internalIdField, parentCfgImplFieldDef, fieldDefs.get(internalIdField.getName()));
-        }
-
-        for (Field schemaField : concat(schemaFields, polymorphicFields)) {
-            addConfigurationImplGetMethod(classDef, schemaField, parentCfgImplFieldDef, fieldDefs.get(fieldName(schemaField)));
-        }
-
-        return classDef;
-    }
-
-    /**
-     * Adds a {@link DynamicConfiguration#specificConfigTree} override for the polymorphic configuration case.
-     *
-     * @param classDef                  Definition of a polymorphic configuration class (parent).
-     * @param schemaClass               Polymorphic configuration schema (parent).
-     * @param polymorphicExtensions     Polymorphic configuration instance schemas (children).
-     * @param polymorphicTypeIdFieldDef Identification field for the polymorphic configuration instance.
-     */
-    private void addCfgSpecificConfigTreeMethod(
-            ClassDefinition classDef,
-            Class<?> schemaClass,
-            Set<Class<?>> polymorphicExtensions,
-            FieldDefinition polymorphicTypeIdFieldDef
-    ) {
-        MethodDefinition specificConfigMtd = classDef.declareMethod(
-                of(PUBLIC),
-                SPECIFIC_CONFIG_TREE_MTD.getName(),
-                type(ConfigurationTree.class)
-        );
-
-        // String tmpStr;
-        Variable tmpStrVar = specificConfigMtd.getScope().createTempVariable(String.class);
-
-        StringSwitchBuilder switchBuilder = new StringSwitchBuilder(specificConfigMtd.getScope())
-                .expression(tmpStrVar)
-                .defaultCase(throwException(ConfigurationWrongPolymorphicTypeIdException.class, tmpStrVar));
-
-        for (Class<?> polymorphicExtension : polymorphicExtensions) {
-            // return new SpecialCfgImpl(this);
-            switchBuilder.addCase(
-                    polymorphicInstanceId(polymorphicExtension),
-                    newInstance(
-                            typeFromJavaClassName(schemasInfo.get(polymorphicExtension).cfgImplClassName),
-                            specificConfigMtd.getThis()
-                    ).ret()
-            );
-        }
-
-        // ConfigNode
-        ParameterizedType nodeType = typeFromJavaClassName(schemasInfo.get(schemaClass).nodeClassName);
-
-        // Object tmpObj;
-        Variable tmpObjVar = specificConfigMtd.getScope().createTempVariable(Object.class);
-
-        // tmpObj = this.refreshValue();
-        // tmpStr = ((ConfigNode) tmpObj).typeId;
-        // switch(tmpStr) ...
-        specificConfigMtd.getBody()
-                .append(tmpObjVar.set(specificConfigMtd.getThis().invoke(REFRESH_VALUE_MTD)))
-                .append(tmpStrVar.set(tmpObjVar.cast(nodeType).getField(polymorphicTypeIdFieldDef.getName(), String.class)))
-                .append(switchBuilder.build())
-                .ret();
-    }
-
-    /**
-     * Adds a {@code DynamicConfiguration#removeMembers} override for the polymorphic configuration case.
-     *
-     * @param classDef                  Definition of a polymorphic configuration class (parent).
-     * @param schemaClass               Polymorphic configuration schema (parent).
-     * @param polymorphicExtensions     Polymorphic configuration instance schemas (children).
-     * @param fieldDefs                 Field definitions for all fields of {@code classDef}.
-     * @param polymorphicFields         Fields of polymorphic extensions.
-     * @param polymorphicTypeIdFieldDef Identification field for the polymorphic configuration instance.
-     */
-    private void addCfgRemoveMembersMethod(
-            ClassDefinition classDef,
-            Class<?> schemaClass,
-            Set<Class<?>> polymorphicExtensions,
-            Map<String, FieldDefinition> fieldDefs,
-            Collection<Field> polymorphicFields,
-            FieldDefinition polymorphicTypeIdFieldDef
-    ) {
-        MethodDefinition removeMembersMtd = classDef.declareMethod(
-                of(PUBLIC),
-                "removeMembers",
-                type(void.class),
-                arg("oldValue", type(Object.class)),
-                arg("members", type(Map.class))
-        );
-
-        // InnerNode oldValue;
-        Variable oldValueVar = removeMembersMtd.getScope().getVariable("oldValue");
-
-        // Map members;
-        Variable membersVar = removeMembersMtd.getScope().getVariable("members");
-
-        // String tmpStr;
-        Variable tmpStrVar = removeMembersMtd.getScope().createTempVariable(String.class);
-
-        StringSwitchBuilder switchBuilder = new StringSwitchBuilder(removeMembersMtd.getScope())
-                .expression(tmpStrVar)
-                .defaultCase(throwException(ConfigurationWrongPolymorphicTypeIdException.class, tmpStrVar));
-
-        for (Class<?> polymorphicExtension : polymorphicExtensions) {
-            Collection<Field> removeFields = polymorphicFields.stream()
-                    .filter(f -> !polymorphicExtension.equals(f.getDeclaringClass()))
-                    .collect(toList());
-
-            BytecodeBlock blockCode = new BytecodeBlock();
-
-            for (Field removeField : removeFields) {
-                // this.removeMember(members, this.field);
-                blockCode
-                        .append(removeMembersMtd.getThis())
-                        .append(membersVar)
-                        .append(getThisFieldCode(removeMembersMtd, fieldDefs.get(fieldName(removeField))))
-                        .invokeVirtual(REMOVE_MEMBER_MTD);
-            }
-
-            switchBuilder.addCase(polymorphicInstanceId(polymorphicExtension), blockCode);
-        }
-
-        // ConfigNode
-        ParameterizedType nodeType = typeFromJavaClassName(schemasInfo.get(schemaClass).nodeClassName);
-
-        // tmpStr = ((ConfigNode) oldValue).typeId;
-        // switch(tmpStr) ...
-        removeMembersMtd.getBody()
-                .append(tmpStrVar.set(oldValueVar.cast(nodeType).getField(polymorphicTypeIdFieldDef.getName(), String.class)))
-                .append(switchBuilder.build())
-                .ret();
-    }
-
-    /**
-     * Adds a {@code DynamicConfiguration#addMembers} override for the polymorphic configuration case.
-     *
-     * @param classDef                  Definition of a polymorphic configuration class (parent).
-     * @param schemaClass               Polymorphic configuration schema (parent).
-     * @param polymorphicExtensions     Polymorphic configuration instance schemas (children).
-     * @param fieldDefs                 Field definitions for all fields of {@code classDef}.
-     * @param polymorphicFields         Fields of polymorphic extensions.
-     * @param polymorphicTypeIdFieldDef Identification field for the polymorphic configuration instance.
-     */
-    private void addCfgAddMembersMethod(
-            ClassDefinition classDef,
-            Class<?> schemaClass,
-            Set<Class<?>> polymorphicExtensions,
-            Map<String, FieldDefinition> fieldDefs,
-            Collection<Field> polymorphicFields,
-            FieldDefinition polymorphicTypeIdFieldDef
-    ) {
-        MethodDefinition removeMembersMtd = classDef.declareMethod(
-                of(PUBLIC),
-                "addMembers",
-                type(void.class),
-                arg("newValue", type(Object.class)),
-                arg("members", type(Map.class))
-        );
-
-        // InnerNode newValue;
-        Variable newValueVar = removeMembersMtd.getScope().getVariable("newValue");
-
-        // Map members;
-        Variable membersVar = removeMembersMtd.getScope().getVariable("members");
-
-        // String tmpStr;
-        Variable tmpStrVar = removeMembersMtd.getScope().createTempVariable(String.class);
-
-        StringSwitchBuilder switchBuilder = new StringSwitchBuilder(removeMembersMtd.getScope())
-                .expression(tmpStrVar)
-                .defaultCase(throwException(ConfigurationWrongPolymorphicTypeIdException.class, tmpStrVar));
-
-        for (Class<?> polymorphicExtension : polymorphicExtensions) {
-            Collection<Field> addFields = polymorphicFields.stream()
-                    .filter(f -> polymorphicExtension.equals(f.getDeclaringClass()))
-                    .collect(toList());
-
-            BytecodeBlock blockCode = new BytecodeBlock();
-
-            for (Field addField : addFields) {
-                // this.addMember(members, this.field);
-                blockCode
-                        .append(removeMembersMtd.getThis())
-                        .append(membersVar)
-                        .append(getThisFieldCode(removeMembersMtd, fieldDefs.get(fieldName(addField))))
-                        .invokeVirtual(ADD_MEMBER_MTD);
-            }
-
-            switchBuilder.addCase(polymorphicInstanceId(polymorphicExtension), blockCode);
-        }
-
-        // ConfigNode
-        ParameterizedType nodeType = typeFromJavaClassName(schemasInfo.get(schemaClass).nodeClassName);
-
-        // tmpStr = ((ConfigNode) newValue).typeId;
-        // switch(tmpStr) ...
-        removeMembersMtd.getBody()
-                .append(tmpStrVar.set(newValueVar.cast(nodeType).getField(polymorphicTypeIdFieldDef.getName(), String.class)))
-                .append(switchBuilder.build())
-                .ret();
-    }
-
 
     /**
      * Creates the bytecode {@code throw new Exception(msg);}.
